window: Correctly get width and height from the WINDOW structure
The maxx and maxy members of ncurses' WINDOW structure don't store
the width and height of the window, but rather the maximum permissible
x and y coordinates. This commit changes the window_new() function to
add one to each of them to reflect the width and height of the window.