cmdbox: Add signals for selection changes
In order to allow the user to change selections, the cmdbox must emit
signals when the start or end of the source or destination selection
have been changed. This commit adds the following four signals that
serve the above purpose:

 - "source_start_changed"
 - "source_end_changed"
 - "destination_start_changed"
 - "destination_end_changed"

A handler for one of these signals must have the following signature:

 int handler(struct cmdbox *box, void *user, struct string *data);

Where `box' is a pointer to the emitting widget, `user' is the pointer
that was passed in `user_data' to widget_add_handler(), and `data' is
a pointer to the string that the user entered into the cmdbox.