editor: Make sure textviews are visible when they have a selection
The postedit textview is supposed to be hidden if it does not have a
selection. Hence, it must be made visible once a selection has been
added. This commit adds a call to widget_set_visible() to the cmdbox
event handlers to make sure the textviews are visible if a selection
has been added.