textview: Clear view before redraw, don't skip update of status line
Textviews are not cleared before being redrawn, causing artifacts to
be visible. Furthermore, because the control flow inside the
_textview_redraw() function is incorrect, the _textview_draw_status()
function will not be called if a user-provided telex could not be
parsed
This commit adds a widget_clear() call to _textview_redraw() and fixes
the control flow so that _textview_draw_status() will always be called.