e: Implement proper commandline parsing, add read-only mode
Files are opened with O_RDWR by default, making it impossible to open
files that the user cannot write to. To allow opening of non-readable
files, this commit adds a read-only mode. Further, this commit adds
proper command line parsing, so that the filename and the mode can be
specified on the command line.