editor: Interpret relative telexes as relative to previous input
This commit makes the editor a bit smarter about the handling of
relative telexes. If the user enters a relative expression, the
editor will attempt to interpret it as relative to the previously
entered expression.
Further, if the user is entering a relative expression for the end
of a selection, the editor will interpret it to be relative to the
start of the selection.