textview: Don't drop characters at the right end of the textview
If a line extends beyond the right end of the textview, the character
that should be printed at the start of the next line is dropped. This
happens because the current column isn't correctly compared with the
textview width, leading to the character to be printed beyond the right
end of the textview.
This commit fixes the comparison, so that no characters are printed
outside of the textview.