editor: Remove preedit/postedit split
The editor has a preedit and postedit textview that was intended
to show the effect of an operation before performing it. The former
textview would display the source selection, while the latter
displays the destination selection.
However, I do not intend to implement operations that read from and
write to the text buffer at the same time, making this split
unnecessary.

This commit removes the postedit buffer and textview.