buffer: Center on start of selection if end was not specified
When a user entered only the start of a selection, the editor shows
the part of the buffer that starts with the selected line. However,
it would be better if the lines above the selection were also shown.

This commit modifies `buffer_get_snippet_telex()` so that it
returns a snippet that is centered on the start of the selection if
no end was specified.