buffer: Add method for overwriting insert
To implement overwriting inserts, the buffer type needs a method
that inserts data into the buffer and overwrites a part of it.

This commit adds the `buffer_overwrite()' method, will overwrite
the buffer as follows: if the caller passed a start and end telex,
the function will only overwrite the destination that is specified
by the start and end telexes. If the insertion is longer than the
destination, the remainder will be a non-overwriting insert.
If the caller specified a start, but not an end, the operation
will overwrite as much of the buffer as necessary. The buffer will
be increased if necessary.