editor: Make it possible to unset selections
The editor does not permit invalid expressions (such as an empty
telex) to be set as the start or end of a selection, making it
impossible to unset a selection.

This commit modifies the editor so that it unsets the selection
when a user presses ^Z or ^X to modify it.