buffer: Add method for erasing text
The buffer type does not have any methods for removing text, which
is necessary for the implementation of an erase function.

This commit adds the `buffer_erase()' method, which erases the part
of a buffer between two telexes, if two telexes where specified. If
the caller specifies only the starting telex, the method will erase
from the specified location until the end of the buffer.