editor: Add callback for cmdbox's `erase_requested' signal
To allow the user to erase text from the buffer, the editor needs
to connect a callback to the cmdbox's erase_requested signal, which
is currently unhandled.

This commit adds a function that erases the user's selection from
the buffer, and connects that function to the `erase_requested'
signal of the cmdbox.