cmdbox: Move code to clear cmdbox into separate method
The `cmdbox_redraw()' method directly calls all ncurses functions
necessary to redraw the command box, without of any intermediate
functions, which makes the method harder to read and refactor.

This commit moves the code to clear the cmdbox into a separate
method `_cmdbox_blank()' which is called from `_cmdbox_redraw()'.