cmdbox: Make cursor position relative to widget position
When the cmdbox widget changes the cursor position, it does not set
the position relative to the widget's position, causing the cursor
to be misplaced when the widget does not start at the left terminal
border.

This commit modifies the `_cmdbox_redraw()' method so that the
cursor is placed relative to the widget's X and Y coordinates.