cmdbox: Store y-coordinate of the cursor
Because the cmdbox widget does not have multiline support, there
was no need to store the y-coordinate of the cursor. To add support
for multiple lines, the y-coordinate needs to be handled as well.

This commit modifies the cmdbox widget so that it stores not just
the cursor's X, but also the Y-coordinate.