editor: Don't intercept ^Q event
The ^Q event is handled by the editor directly, without forwarding
it to other widgets, making it impossible to configure and handle
all keyboard shortcuts in a unified way.

This commit modifies the `editor_run()' method so that it does not
handle ^Q events but instead forwards them to the window.