kbdwidget: Add widget that implements handling of key strokes
Low-level handling of key strokes is implemented directly by the
cmdbox widget, where it cannot be reused. This code needs to be
separated so that it can be reused by other widgets.

This commit adds the `kbd_widget' type, which implements low-level
handling of keystrokes, providing a clean interface to base other
widgets upon.