cmdbox: Add viewport that moves with the cursor
The cmdbox widget does not allow the user to insert more characters
than the widget can display, making it unable to accept long inputs.

This commit adds a viewport to the widget that moves with the cursor,
so that the widget always displays the input near the cursor, like
one is used to from GtkEntry widgets and such.