editor: Also unset selection in editor
When unsetting the start or end of the selection, the selection is
unset in the textview, causing the view to be updated, but it is
not unset in the editor itself. Thus, the selection is still used
when an action is performed, even though it appears to be unset.

This commit modifies the editor so that unsetting the start or end
of a selection also causes it to be unset in the editor.