buffer: Tell caller where insertions ended
It would be more user-friendly if the editor automatically moved
the selection when the user inserted some text. However, to move
the selection, the editor needs to know where an insertion ended.

This commit modifies the `buffer_insert()' and `buffer_overwrite()'
functions so that they take an (optional) argument `new_end', which
the functions will adjust to point to the end of the insertion. A
caller that does not care where an insertion ends, may pass NULL.