editor: Advance selection when user performed an insertion
When the user makes an insertion, the editor does not advance the
selection. Thus, the user has to manually advance the selection if
they want to continue inserting text.

This commit modifies the editor so that it sets the start of the
selection to the position where the last insertion ended.