editor: Remove handler for `size_changed' signal of cmdbox widget
When initializing an editor object, `_editor_init_ui()' attempts
to register a callback to the non-existing signal `size_changed'
of the editor's command box, causing the initialization to fail.

This commit removes the handler for the invalid signal and removes
the call to `widget_add_handler()' that attempts to register the
handler.