window: Don't access members of WINDOW structure directly
When determining the size of the terminal, the `window_new()' and
`_window_resize()' methods access members of the WINDOW structure
directly. In more recent versions of ncurses, WINDOW is an opaque
structure, causing errors at compile time.

This commit modifies the `window_new()' and `_window_resize()'
methods so they use `getmaxyx()' to determine the size of the
terminal.