multistring: Add data type for backing multi-line cmdbox
The cmdbox widget uses a string to store the user input. To add
multi-line support, the cmdbox thus has to fiddle with the content
of the string directly, making the implementation more complicated.
To separate the string-handling from the widget, a new data type is
required.

This commit adds the multistring data type, which provides an API
similar to that of the string type, but with support for multiple
lines.