editor: Add support for multi-line cmdbox
The editor widget expects the cmdbox widget to use a string for
storing the user input, but multi-line cmdbox widgets use a
multistring. This causes the editor widget to use string methods
on a multistring type.

This commit modifies the editor widget so that it uses multistring
methods to handle the cmdbox's data.