cmdbox: Correctly handle tabs
The cmdbox widget does not handle tabs correctly. Neither are tab
characters inserted into the input buffer when the user hits TAB or
CTRL + I, nor are tabs displayed correctly when they are contained
in the input buffer.

This commit modifies the cmdbox widget so that it inserts a '\t'
character into the input buffer when the TAB or CTRL + I keys are
pressed. This commit further modifies the `_cmdbox_redraw()' method
so that the appropriate amount of spaces are written to the screen
where a tab is in the input buffer, and the position of the cursor
is adjusted accordingly.