buffer: Skip saving if buffer is unmodified
The `buffer_save()' method will save the buffer to the file system
even if the buffer has not been modified since it was last saved or
since it was read from the file system. Thus, if the user hits the
keystroke for saving the buffer twice in a row, the buffer will be
written twice, causing unnecessary work for the file system and
storage devices.

This commit modifies the buffer so that it tracks whether it has
been modified, allowing the `buffer_save()' method to skip writing
unmodified buffers to the file system.