buffer: Overwrite entire selection from start to end
When `buffer_overwrite()' is called with a valid selection start
and end, but the insertion is smaller than the selection, the
function will overwrite only a part of the selection. However, it
would be more intuitive if the method would remove the part of the
selection that was not overwritten, effectively replacing the
selection with the insertion.

This commit modifies the `buffer_overwrite()' function so that it
will replace the selection with the insertion if both the start and
the end of the selection were specified.